feat(TransactionSyncWorker): moved handler to specific file
