+++ src/plugins/Workers/TransactionSyncStreamWorker/handlers/onStreamError.js
@@ -1,0 +1,7 @@
+const logger = require('../../../../logger');
+
+function onStreamError(error, reject) {
+  logger.silly('TransactionSyncStreamWorker - end stream on error');
+  reject(error);
+}
+module.exports = onStreamError;
